More documentation and slight refactor. (Unsure if engine starts atm due to unfinished Script changes)

+++ CamelotUtility/Include/CmMemAllocProfiler.h
@@ -4,6 +4,7 @@ namespace CamelotFramework
 {
 	/**
 	 * @brief	Specialized allocator for profiler so we can avoid tracking internal profiler memory allocations
+	 * 			which would skew profiler results.
 	 */
 	class ProfilerAlloc
 	{};
@@ -16,21 +17,34 @@ namespace CamelotFramework
 	class MemoryAllocator<ProfilerAlloc> : public MemoryAllocatorBase
 	{
 	public:
-		static inline void* allocate(UINT32 bytes)
+		/**
+		 * @brief	Allocates the given number of bytes.
+		 */
+		static inline void* allocate(size_t bytes)
 		{
 			return malloc(bytes);
 		}
 
-		static inline void* allocateArray(UINT32 bytes, UINT32 count)
+		/**
+		 * @brief	Allocates the given a number of objects, each of the given number of bytes.
+		 */
+		static inline void* allocateArray(size_t bytes, UINT32 count)
 		{
 			return malloc(bytes * count);
 		}
 
+		/**
+		 * @brief	Frees memory previously allocated with "allocate".
+		 */
 		static inline void free(void* ptr)
 		{
 			::free(ptr);
 		}
 
+		/**
+		 * @brief	Frees memory previously allocated with "freeArray". "count" must match the
+		 * 			original value when array was allocated.
+		 */
 		static inline void freeArray(void* ptr, UINT32 count)
 		{
 			::free(ptr);
